Develop

+++ src/main/java/hu/bme/edu/handmade/controllers/UserController.java
@@ -1,28 +1,34 @@
 package hu.bme.edu.handmade.controllers;
-
+import hu.bme.edu.handmade.exception.ForbiddenException;
 import hu.bme.edu.handmade.mappers.AddressMapper;
 import hu.bme.edu.handmade.mappers.UserMapper;
 import hu.bme.edu.handmade.models.CartProduct;
 import hu.bme.edu.handmade.models.User;
+import hu.bme.edu.handmade.services.IEmailService;
 import hu.bme.edu.handmade.services.IUserService;
 import hu.bme.edu.handmade.services.IWishListService;
 import hu.bme.edu.handmade.services.impl.CartProductService;
 import hu.bme.edu.handmade.web.dto.CartProductDto;
+import hu.bme.edu.handmade.web.dto.PasswordDto;
 import hu.bme.edu.handmade.web.dto.ProductDto;
 import hu.bme.edu.handmade.web.dto.user.AddressDto;
+import hu.bme.edu.handmade.web.dto.user.RoleDto;
 import hu.bme.edu.handmade.web.dto.user.UserDto;
-import hu.bme.edu.handmade.web.dto.error.ResourceNotFoundException;
+import hu.bme.edu.handmade.exception.ResourceNotFoundException;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
-
 import java.net.URI;
+import java.net.URLDecoder;
+import java.nio.charset.StandardCharsets;
 import java.security.Principal;
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
+import java.util.UUID;
 
 @RestController
 @CrossOrigin(origins="http://localhost:4200", maxAge=3600)
@@ -31,11 +37,13 @@ public class UserController {
     private final IUserService userService;
     private final CartProductService cartService;
     private final IWishListService wishListService;
+    private final IEmailService emailService;
 
-    UserController(IUserService userService, CartProductService cartService, IWishListService wishListService) {
+    UserController(IUserService userService, CartProductService cartService, IWishListService wishListService, IEmailService emailService) {
         this.userService = userService;
         this.cartService = cartService;
         this.wishListService = wishListService;
+        this.emailService = emailService;
     }
 
     @PreAuthorize("hasAuthority('ROLE_USER')")
@@ -44,15 +52,6 @@ public class UserController {
     public UserDto user(Principal principal) {
         String name = principal.getName();
         return UserMapper.INSTANCE.userToUserDto(userService.findUserByEmail(name));
-    }
-
-    @PreAuthorize("hasAuthority('ROLE_ADMIN')")
-    @SecurityRequirement(name = "Bearer_Authentication")
-    @GetMapping("/{id}")
-    public UserDto one(@PathVariable Long id) {
-        User u = userService.getUserByID(id)
-                .orElseThrow(() -> new ResourceNotFoundException("User with id: " + id + " is not found."));
-        return UserMapper.INSTANCE.userToUserDto(u);
     }
 
     @SecurityRequirement(name = "Bearer_Authentication")
@@ -75,10 +74,28 @@ public class UserController {
         }
     }
 
+    @SecurityRequirement(name = "Bearer_Authentication")
+    @PreAuthorize("hasAuthority('ROLE_USER')")
+    @DeleteMapping(path = { "/delete" })
+    public ResponseEntity<?> deleteOwnProfile(Principal principal) {
+        User user = userService.findUserByEmail(principal.getName());
+        try {
+            userService.deleteUser(user.getId());
+            return ResponseEntity.noContent().build();
+        }
+        catch (EmptyResultDataAccessException e) {
+            return ResponseEntity.notFound().build();
+        }
+    }
+
     @PreAuthorize("hasAuthority('ROLE_USER')")
     @SecurityRequirement(name = "Bearer_Authentication")
     @PutMapping("/{id}")
-    public ResponseEntity<User> updateUser(@PathVariable("id") Long id, @RequestBody UserDto userToUpdate) {
+    public ResponseEntity<User> updateUser(@PathVariable("id") Long id, @RequestBody UserDto userToUpdate, Principal principal) {
+        User principalUser = userService.findUserByEmail(principal.getName());
+        if (!Objects.equals(principalUser.getId(), id)) {
+            throw new ForbiddenException();
+        }
         Optional<User> updatedUser = userService.updateUser(id, userToUpdate);
 
         return updatedUser
@@ -95,36 +112,74 @@ public class UserController {
     @PreAuthorize("hasAuthority('ROLE_USER')")
     @SecurityRequirement(name = "Bearer_Authentication")
     @PostMapping("/{id}/addresses")
-    public AddressDto addUserAddress(@PathVariable("id") Long userId, @RequestBody AddressDto address) {
+    public AddressDto addUserAddress(@PathVariable("id") Long userId, @RequestBody AddressDto address, Principal principal) {
+        User principalUser = userService.findUserByEmail(principal.getName());
+        if (!Objects.equals(principalUser.getId(), userId)) {
+            throw new ForbiddenException();
+        }
         return AddressMapper.INSTANCE.toAddressDto(userService.addAddress(userId, address));
     }
 
     @PreAuthorize("hasAuthority('ROLE_USER')")
     @SecurityRequirement(name = "Bearer_Authentication")
     @PutMapping("/{id}/addresses/{addressId}")
-    public AddressDto updateUserAddress(@PathVariable("id") Long userId, @PathVariable("addressId") Long addressId,@RequestBody AddressDto address) {
+    public AddressDto updateUserAddress(@PathVariable("id") Long userId,
+                                        @PathVariable("addressId") Long addressId,
+                                        @RequestBody AddressDto address,
+                                        Principal principal) {
+        User principalUser = userService.findUserByEmail(principal.getName());
+        if (!Objects.equals(principalUser.getId(), userId)) {
+            throw new ForbiddenException();
+        }
         return AddressMapper.INSTANCE.toAddressDto(userService.updateAddress(userId, address));
+    }
+
+    /** Newsletter subscription */
+    @PreAuthorize("hasAuthority('ROLE_USER')")
+    @SecurityRequirement(name = "Bearer_Authentication")
+    @PostMapping("/{id}/unsubscribe")
+    public void unsubscribeFromNewsletter(@PathVariable("id") Long userId, Principal principal) {
+        User principalUser = userService.findUserByEmail(principal.getName());
+        if (!Objects.equals(principalUser.getId(), userId)) {
+            throw new ForbiddenException();
+        }
+        userService.setNewsletterSubscription(userId, false);
     }
 
     /** Cart */
     @PreAuthorize("hasAuthority('ROLE_USER')")
     @SecurityRequirement(name = "Bearer_Authentication")
     @GetMapping("/{id}/cart")
-    List<CartProduct> getCartProducts(@PathVariable("id") long userId) {
+    List<CartProduct> getCartProducts(@PathVariable("id") long userId, Principal principal) {
+        User principalUser = userService.findUserByEmail(principal.getName());
+        if (!Objects.equals(principalUser.getId(), userId)) {
+            throw new ForbiddenException();
+        }
         return cartService.getCartProductsByUser(userId);
     }
 
     @PreAuthorize("hasAuthority('ROLE_USER')")
     @SecurityRequirement(name = "Bearer_Authentication")
     @PostMapping("/{id}/cart")
-    CartProduct addCartProduct(@PathVariable("id") long userId, @RequestBody CartProductDto cartProductDto) {
+    CartProduct addCartProduct(@PathVariable("id") long userId, @RequestBody CartProductDto cartProductDto, Principal principal) {
+        User principalUser = userService.findUserByEmail(principal.getName());
+        if (!Objects.equals(principalUser.getId(), userId)) {
+            throw new ForbiddenException();
+        }
         return cartService.addCartProduct(cartProductDto, userId);
     }
 
     @PreAuthorize("hasAuthority('ROLE_USER')")
     @SecurityRequirement(name = "Bearer_Authentication")
     @PutMapping("/{userId}/cart/{cartId}")
-    CartProduct updateCartProduct(@PathVariable("userId") long userId, @PathVariable("cartId") long cartId, @RequestBody CartProductDto cartProductDto) {
+    CartProduct updateCartProduct(@PathVariable("userId") long userId,
+                                  @PathVariable("cartId") long cartId,
+                                  @RequestBody CartProductDto cartProductDto,
+                                  Principal principal) {
+        User principalUser = userService.findUserByEmail(principal.getName());
+        if (!Objects.equals(principalUser.getId(), userId)) {
+            throw new ForbiddenException();
+        }
         return cartService.findCartProductById(cartId)
                 .map(product -> cartService.updateCartProduct(cartProductDto, cartId))
                 .orElseGet(()->cartService.addCartProduct(cartProductDto, userId));
@@ -133,7 +188,13 @@ public class UserController {
     @PreAuthorize("hasAuthority('ROLE_USER')")
     @SecurityRequirement(name = "Bearer_Authentication")
     @DeleteMapping("/{userId}/cart/{cartId}")
-    ResponseEntity<?> removeCartProduct(@PathVariable("userId") Long userId, @PathVariable("cartId") Long cartId) {
+    ResponseEntity<?> removeCartProduct(@PathVariable("userId") Long userId,
+                                        @PathVariable("cartId") Long cartId,
+                                        Principal principal) {
+        User principalUser = userService.findUserByEmail(principal.getName());
+        if (!Objects.equals(principalUser.getId(), userId)) {
+            throw new ForbiddenException();
+        }
         try {
             cartService.deleteCartProduct(cartId);
             return ResponseEntity.noContent().build();
@@ -147,21 +208,81 @@ public class UserController {
     @PreAuthorize("hasAuthority('ROLE_USER')")
     @SecurityRequirement(name = "Bearer_Authentication")
     @GetMapping("/{user_id}/wishlist")
-    List<ProductDto> getWishList(@PathVariable("user_id") Long id) {
+    List<ProductDto> getWishList(@PathVariable("user_id") Long id, Principal principal) {
+        User principalUser = userService.findUserByEmail(principal.getName());
+        if (!Objects.equals(principalUser.getId(), id)) {
+            throw new ForbiddenException();
+        }
         return wishListService.getWishListItemsByUserId(id);
     }
 
     @PreAuthorize("hasAuthority('ROLE_USER')")
     @SecurityRequirement(name = "Bearer_Authentication")
     @PostMapping("/{user_id}/wishlist")
-    ProductDto addToWishList(@PathVariable("user_id") Long userId, @RequestParam("product_id") Long productId) {
+    ProductDto addToWishList(@PathVariable("user_id") Long userId,
+                             @RequestParam("product_id") Long productId,
+                             Principal principal) {
+        User principalUser = userService.findUserByEmail(principal.getName());
+        if (!Objects.equals(principalUser.getId(), userId)) {
+            throw new ForbiddenException();
+        }
         return wishListService.addToWishlist(userId, productId);
     }
 
     @PreAuthorize("hasAuthority('ROLE_USER')")
     @SecurityRequirement(name = "Bearer_Authentication")
     @DeleteMapping("/{user_id}/wishlist")
-    void removeFromWishList(@PathVariable("user_id") Long userId, @RequestParam("product_id") Long productId) {
+    void removeFromWishList(@PathVariable("user_id") Long userId,
+                            @RequestParam("product_id") Long productId,
+                            Principal principal) {
+        User principalUser = userService.findUserByEmail(principal.getName());
+        if (!Objects.equals(principalUser.getId(), userId)) {
+            throw new ForbiddenException();
+        }
         wishListService.removeFromWishList(userId, productId);
+    }
+
+    /** Reset Password */
+    @PostMapping("/sendPswResetRequest")
+    public ResponseEntity<?> sendResetPasswordMail(@RequestParam("email") String userEmail) {
+        String decodedEmail = URLDecoder.decode(userEmail, StandardCharsets.UTF_8);
+        User user = userService.findUserByEmail(decodedEmail);
+        if (user == null) {
+            throw new ResourceNotFoundException("User with email: " + userEmail + " is not found.");
+        }
+        String token = UUID.randomUUID().toString();
+        userService.createPasswordResetTokenForUser(user, token);
+        return emailService.sendPasswordReset(decodedEmail, token);
+    }
+
+    @PostMapping("/updatePassword")
+    public ResponseEntity<User> updatePassword(PasswordDto passwordDto) {
+        String tokenError = userService.validatePasswordResetToken(passwordDto.getToken());
+        if (tokenError != null) {
+            return ResponseEntity.badRequest().build();
+        }
+        User user = userService.getUserByPasswordResetToken(passwordDto.getToken());
+        User updatedUser = userService.changeUserPassword(user, passwordDto.getNewPassword());
+        return ResponseEntity.ok(updatedUser);
+    }
+
+    @PreAuthorize("hasAuthority('ROLE_ADMIN')")
+    @SecurityRequirement(name = "Bearer_Authentication")
+    @PostMapping("/setRole")
+    public void setRole(@RequestParam("userId") Long userId, @RequestParam("roleIds") List<Long> roleIds) {
+        userService.setRoles(userId, roleIds);
+    }
+
+    @PreAuthorize("hasAuthority('ROLE_ADMIN')")
+    @SecurityRequirement(name = "Bearer_Authentication")
+    @GetMapping("/roles")
+    public List<RoleDto> getRoles() {
+        return userService.getAllRoles();
+    }
+    @PreAuthorize("hasAuthority('ROLE_ADMIN')")
+    @SecurityRequirement(name = "Bearer_Authentication")
+    @GetMapping("/{userId}/roles")
+    public List<RoleDto> getUserRoles(@PathVariable("userId") Long userId) {
+        return userService.getUserRoles(userId);
     }
 }
